added config for disabling direct messages off by default; moved mutation for sending message to user to comms module; removed retrieval of communityRooms for a user
