+++ alkemio.yml
@@ -284,6 +284,9 @@ communications:
     connection_timeout: ${COMMUNICATIONS_MATRIX_CONNECTION_TIMEOUT}:30
   discussions:
     enabled: ${COMMUNICATIONS_DISCUSSIONS_ENABLED}:true
+  direct_message_rooms:
+    enabled: ${COMMUNICATIONS_DIRECT_MESSAGE_ROOMS_ENABLED}:false
+
 
 ## storage ##
 # Alkemio uses multiple types of persistent storage, including SQL database, postgres database, file storage, redis.
