+++ package.json
@@ -150,6 +150,7 @@
     "jest-mock": "^29.7.0",
     "lint-staged": "^15.2.7",
     "prettier": "^3.3.3",
+    "rimraf": "^5.0.0",
     "supertest": "^7.0.0",
     "ts-jest": "^29.2.2",
     "ts-node": "^10.9.2",
