+++ src/domain/communication/communication/communication.resolver.mutations.ts
@@ -20,12 +20,15 @@ import { NotificationUserAdapter } from '@services/adapters/notification-adapter
 import { UserService } from '@domain/community/user/user.service';
 import { MessagingNotEnabledException } from '@common/exceptions/messaging.not.enabled.exception';
 import { LogContext } from '@common/enums/logging.context';
+import { CommunicationAdapter } from '@services/adapters/communication-adapter/communication.adapter';
+import { UserSendMessageInput } from '@domain/communication/communication/dto/user.dto.communication.message.send';
 
 @InstrumentResolver()
 @Resolver()
 export class CommunicationResolverMutations {
   constructor(
     private authorizationService: AuthorizationService,
+    private communicationAdapter: CommunicationAdapter,
     private notificationAdapterSpace: NotificationSpaceAdapter,
     private notificationUserAdapter: NotificationUserAdapter,
     private notificationOrganizationAdapter: NotificationOrganizationAdapter,
@@ -33,6 +36,48 @@ export class CommunicationResolverMutations {
     private userService: UserService,
     @Inject(WINSTON_MODULE_NEST_PROVIDER) private readonly logger: LoggerService
   ) {}
+
+  @Mutation(() => String, {
+    description:
+      'Sends a message on the specified User`s behalf and returns the room id',
+  })
+  async messageUser(
+    @Args('messageData') messageData: UserSendMessageInput,
+    @CurrentUser() agentInfo: AgentInfo
+  ): Promise<string> {
+    const receivingUser = await this.userService.getUserOrFail(
+      messageData.receivingUserID,
+      {
+        relations: {
+          settings: true,
+        },
+      }
+    );
+    this.authorizationService.grantAccessOrFail(
+      agentInfo,
+      receivingUser.authorization,
+      AuthorizationPrivilege.READ,
+      `user send message: ${receivingUser.id}`
+    );
+
+    // Check if the user is willing to receive messages
+    if (!receivingUser.settings.communication.allowOtherUsersToSendMessages) {
+      throw new MessagingNotEnabledException(
+        'User is not open to receiving messages',
+        LogContext.USER,
+        {
+          userId: receivingUser.id,
+          senderId: agentInfo.userID,
+        }
+      );
+    }
+
+    return await this.communicationAdapter.sendMessageToUser({
+      senderCommunicationsID: agentInfo.communicationID,
+      message: messageData.message,
+      receiverCommunicationsID: receivingUser.communicationID,
+    });
+  }
 
   @Mutation(() => Boolean, {
     description: 'Send message to a User.',
