+++ src/domain/communication/communication/dto/user.dto.communication.message.send.ts
