+++ src/domain/communication/communication/dto/user.dto.communication.room.direct.result.ts
