+++ src/domain/community/user/dto/index.ts
@@ -1,4 +1,4 @@
 export * from './user.dto.create';
 export * from './user.dto.update';
 export * from './user.dto.delete';
-export * from './user.dto.communication.message.send';
+export * from '../../../communication/communication/dto/user.dto.communication.message.send';
