+++ src/domain/community/user/user.resolver.mutations.ts
@@ -1,29 +1,24 @@
 import { Inject, LoggerService } from '@nestjs/common';
 import { Args, Mutation, Resolver } from '@nestjs/graphql';
-import { CurrentUser, Profiling } from '@src/common/decorators';
+import { CurrentUser } from '@src/common/decorators';
 import { IUser } from '@domain/community/user/user.interface';
 import { UserService } from './user.service';
 import { AuthorizationService } from '@core/authorization/authorization.service';
 import { AuthorizationPrivilege } from '@common/enums/authorization.privilege';
 import { AgentInfo } from '@core/authentication.agent.info/agent.info';
 import { UserAuthorizationService } from './user.service.authorization';
-import { UserSendMessageInput } from './dto/user.dto.communication.message.send';
 import { UserAuthorizationResetInput } from './dto/user.dto.reset.authorization';
-import { CommunicationAdapter } from '@services/adapters/communication-adapter/communication.adapter';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { UpdateUserPlatformSettingsInput } from './dto/user.dto.update.platform.settings';
 import { UpdateUserInput } from './dto';
 import { AuthorizationPolicyService } from '@domain/common/authorization-policy/authorization.policy.service';
 import { UpdateUserSettingsInput } from './dto/user.dto.update.settings';
 import { InstrumentResolver } from '@src/apm/decorators';
-import { LogContext } from '@common/enums';
-import { MessagingNotEnabledException } from '@common/exceptions/messaging.not.enabled.exception';
 
 @InstrumentResolver()
 @Resolver(() => IUser)
 export class UserResolverMutations {
   constructor(
-    private communicationAdapter: CommunicationAdapter,
     private authorizationService: AuthorizationService,
     private authorizationPolicyService: AuthorizationPolicyService,
     private userService: UserService,
@@ -83,52 +78,9 @@ export class UserResolverMutations {
     return this.userService.getUserOrFail(user.id);
   }
 
-  @Mutation(() => String, {
-    description:
-      'Sends a message on the specified User`s behalf and returns the room id',
-  })
-  async messageUser(
-    @Args('messageData') messageData: UserSendMessageInput,
-    @CurrentUser() agentInfo: AgentInfo
-  ): Promise<string> {
-    const receivingUser = await this.userService.getUserOrFail(
-      messageData.receivingUserID,
-      {
-        relations: {
-          settings: true,
-        },
-      }
-    );
-    this.authorizationService.grantAccessOrFail(
-      agentInfo,
-      receivingUser.authorization,
-      AuthorizationPrivilege.READ,
-      `user send message: ${receivingUser.id}`
-    );
-
-    // Check if the user is willing to receive messages
-    if (!receivingUser.settings.communication.allowOtherUsersToSendMessages) {
-      throw new MessagingNotEnabledException(
-        'User is not open to receiving messages',
-        LogContext.USER,
-        {
-          userId: receivingUser.id,
-          senderId: agentInfo.userID,
-        }
-      );
-    }
-
-    return await this.communicationAdapter.sendMessageToUser({
-      senderCommunicationsID: agentInfo.communicationID,
-      message: messageData.message,
-      receiverCommunicationsID: receivingUser.communicationID,
-    });
-  }
-
   @Mutation(() => IUser, {
     description: 'Reset the Authorization policy on the specified User.',
   })
-  @Profiling.api
   async authorizationPolicyResetOnUser(
     @CurrentUser() agentInfo: AgentInfo,
     @Args('authorizationResetData')
